Change how types are handled
