+++ src/graphql/mutation/createProcedureMutationField.js
@@ -1,0 +1,3 @@
+const createProcedureMutationField = procedure => ({
+  table: getType()
+})
